File management, efficiency and ramps
I made a folder for the images for better file management.
To make the code more efficient, I deleted all of those if statements and made it an array and arcade would then spawn the item in the array that's in the position that the .csv specifies. It is important to remember that arrays start at 0 for this to work.
I made the ramps work by editing the collision points, essentially halving them.